+++ osa6/redux-anecdotes/src/App.jsx
@@ -1,11 +1,13 @@
 import AnecdoteList from './components/AnecdoteList'
 import AnecdoteForm from './components/AnecdoteForm'
 import Filter from './components/Filter'
+import Notification from './components/Notification'
 
 const App = () => {
   return (
     <div>
       <h2>Anecdotes</h2>
+      <Notification />
       <Filter />
       <AnecdoteList />
       <AnecdoteForm />
