+++ osa6/redux-anecdotes/src/reducers/notificationReducer.js
@@ -1,0 +1,16 @@
+import { createSlice } from '@reduxjs/toolkit'
+
+const initialState = 'initial notification'
+
+const notificationSlice = createSlice({
+  name: 'notification',
+  initialState,
+  reducers: {
+    setNotification(state, action) {
+      return action.payload
+    }
+  }
+})
+
+export const { setNotification } = notificationSlice.actions
+export default notificationSlice.reducer
