+++ osa6/redux-anecdotes/src/utils/store.js
@@ -1,11 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit'
 import anecdoteReducer from '../reducers/anecdoteReducer'
 import filterReducer from '../reducers/filterReducer'
+import notificationReducer from '../reducers/notificationReducer'
 
 const store = configureStore({
   reducer: {
     anecdotes: anecdoteReducer,
-    filter: filterReducer
+    filter: filterReducer,
+    notification: notificationReducer
   }, devTools: true
 })
